// src/app/core/api/v1/model/meetingSession.ts
/**
 * Quarkus CIS Home service
 * Quarkus ChristInformationSystem
 *
 * The version of the OpenAPI document: 1.0
 * Contact: dominik@christ-rlp.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeetingDate } from './meetingDate';
import { Participant } from './participant';
import { MeetingDescription } from './meetingDescription';


export interface MeetingSession { 
    meetingSessionId?: string;
    meeting?: string;
    meetingDate?: MeetingDate;
    participantList?: Array<Participant>;
    description?: MeetingDescription;
    meetingName?: string;
    meetingId?: string;
    participants?: Array<Participant>;
}